fixed: 본인 게시글만 수정,삭제 가능하게함, 상세조회에서 찜갯수 포함

+++ src/main/java/com/sparta/backend/service/PostService.java
@@ -49,7 +49,18 @@ public class PostService {
         for (Post post : postList) {
             Long dibCount = dibsRepository.countByPost(post);
             Long commentsCount = commentRepository.countByPost(post);
-            allPostResponseDtoList.add(AllPostResponseDto.builder().id(post.getId()).title(post.getTitle()).content(post.getContent()).imgUrl(post.getImgUrl()).price(post.getPrice()).category(post.getCategory()).dibCount(dibCount).view(post.getViews()).commentsCount(commentsCount).createdAt(post.getCreatedAt()).modifiedAt(post.getModifiedAt()).build());
+            allPostResponseDtoList.add(AllPostResponseDto.builder()
+                    .id(post.getId())
+                    .title(post.getTitle())
+                    .content(post.getContent())
+                    .imgUrl(post.getImgUrl())
+                    .price(post.getPrice())
+                    .category(post.getCategory())
+                    .dibCount(dibCount)
+                    .view(post.getViews())
+                    .commentsCount(commentsCount)
+                    .createdAt(post.getCreatedAt())
+                    .modifiedAt(post.getModifiedAt()).build());
         }
         return ResponseDto.success(allPostResponseDtoList);
     }
@@ -82,8 +93,16 @@ public class PostService {
         addViewCount(post.get());
         List<Comment> commentList = commentRepository.findAllByPost(post.get());
         List<CommentResponseDto> commentResponseDtoList = new ArrayList<>();
+        Long dibCount = dibsRepository.countByPost(post.get());
         for (Comment comment : commentList) {
-            commentResponseDtoList.add(CommentResponseDto.builder().id(comment.getId()).memberId(comment.getMember().getMemberId()).content(comment.getContent()).createdAt(comment.getCreatedAt()).modifiedAt(comment.getModifiedAt()).postId(comment.getPost().getId()).build());
+            commentResponseDtoList.add(CommentResponseDto.builder()
+                    .id(comment.getId())
+                    .memberId(comment.getMember().getMemberId())
+                    .content(comment.getContent())
+                    .createdAt(comment.getCreatedAt())
+                    .modifiedAt(comment.getModifiedAt())
+                    .postId(comment.getPost().getId())
+                    .build());
         }
         return ResponseDto.success(PostResponseDto.builder()
                 .id(post.get().getId())
@@ -94,9 +113,9 @@ public class PostService {
                 .category(post.get().getCategory())
                 .views(post.get().getViews())
                 .commentResponseDtoList(commentResponseDtoList)
+                .dibCount(dibCount)
                 .memberId(post.get().getMember().getMemberId())
                 .build());
-
     }
 
     @Transactional
@@ -114,7 +133,7 @@ public class PostService {
         }
         boolean dibsResult;
         dibsResult = null != dibsRepository.findByMemberAndPost(member, post.get());
-        return  ResponseDto.success(dibsResult);
+        return ResponseDto.success(dibsResult);
     }
 
     @Transactional
@@ -152,6 +171,11 @@ public class PostService {
         if (post.isEmpty()) {
             throw new NotFoundException("게시글을 찾을 수 없습니다.");
         }
+
+        if (post.get().validateMember(member)) {
+            return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
+        }
+
         String key = post.get().getImgUrl().substring("https://mysparta00.s3.ap-northeast-2.amazonaws.com/".length());
         amazonS3Client.deleteObject(bucket, key);
 
@@ -177,6 +201,11 @@ public class PostService {
         if (post.isEmpty()) {
             throw new NotFoundException("게시글을 찾을 수 없습니다.");
         }
+
+        if (post.get().validateMember(member)) {
+            return ResponseDto.fail("BAD_REQUEST", "작성자만 수정할 수 있습니다.");
+        }
+
         String key = post.get().getImgUrl().substring("https://mysparta00.s3.ap-northeast-2.amazonaws.com/".length());
         amazonS3Client.deleteObject(bucket, key);
         postRepository.delete(post.get());
@@ -192,7 +221,19 @@ public class PostService {
         for (Post post : postList) {
             Long dibCount = dibsRepository.countByPost(post);
             Long commentsCount = commentRepository.countByPost(post);
-            categoryList.add(AllPostResponseDto.builder().id(post.getId()).title(post.getTitle()).content(post.getContent()).imgUrl(post.getImgUrl()).price(post.getPrice()).category(post.getCategory()).dibCount(dibCount).view(post.getViews()).commentsCount(commentsCount).createdAt(post.getCreatedAt()).modifiedAt(post.getModifiedAt()).build());
+            categoryList.add(AllPostResponseDto.builder()
+                    .id(post.getId())
+                    .title(post.getTitle())
+                    .content(post.getContent())
+                    .imgUrl(post.getImgUrl())
+                    .price(post.getPrice())
+                    .category(post.getCategory())
+                    .dibCount(dibCount)
+                    .view(post.getViews())
+                    .commentsCount(commentsCount)
+                    .createdAt(post.getCreatedAt())
+                    .modifiedAt(post.getModifiedAt())
+                    .build());
         }
         return ResponseDto.success(categoryList);
     }
